break out step size
